Adding Count Endpoint Code
Broke out Filter into it's own file class. Adding code to enable count endpoint optimization.
